refactor: Phase 3 slice 3e — skip-detector router + auto-skip orchestration into SkipAnalysisService (#84)
Fifth slice (3e) of Phase 3 — migrates the 9 skip-detector handlers
plus DOWNLOAD_INJECT_CHAPTERS behind the router pattern, and folds the
auto-skip orchestration that Phase 2 PR4 deferred into the
SkipAnalysisService. Behavior-identical, no IPC surface change.

* src/main/services/skip-analysis/index.ts grows from the Phase 2
  leaf surface to also own the orchestration that carried in-flight
  state:
  - full-show analysis single-flight: runSkipAnalysis,
    getCurrentAnalysis, cancelCurrentAnalysis
  - per-renderer stream detection: runStreamSkipDetection,
    cancelStreamSkipDetection
  - debounced auto-skip queue: scheduleAutoSkipAnalysis,
    enqueueAutoSkipAnalysis, getAutoSkipQueueStatus
  All in-flight state (currentSkipAnalysis, the stream-detection map,
  the debounce map + queue + draining flag) now lives in the service
  closure. New deps: analyzeProgressChannel, getFpcalcPath,
  getFfmpegPath, getFfprobePath; the service imports analyzeShow /
  detectStream directly.

* New src/main/ipc/skip-detector.ipc.ts — CHANNELS.SKIP_DETECTOR_
  {GET_DETECTIONS, GET_STATUS, CANCEL, CACHE_STATS, ANALYZE_SHOW,
  DETECT_STREAM, CANCEL_STREAM_DETECT, BACKFILL_ALL, QUEUE_STATUS}
  plus DOWNLOAD_INJECT_CHAPTERS (deferred from 3c — it depends on the
  skip-analysis single-flight). Chapter injection still ffprobe-gates,
  writes via fluent-ffmpeg -map_chapters, and rewrites the fingerprint
  cache key when the file size/mtime changes.

* src/main/index.ts removes ~185 LOC of skip orchestration + ~210 LOC
  of handlers; rewires the two external call sites (the window
  destroyed hook → cancelStreamSkipDetection, the download/merge
  completion hooks → scheduleAutoSkipAnalysis) to the service; drops
  the now-unused os / analyzeShow / detectStream / fingerprintCacheKey
  / formatChaptersMetadata / EpisodeInput / AnalyzeProgress /
  EpisodeSkipDetection imports.

* test/services/skip-analysis.test.ts updated for the 4 new
  constructor deps (existing 13 tests still pass).

* registerIpcHandlers() down to ~36 handlers (settings,
  watch-progress, storage, file/shell, player, syncplay,
  auto-download, debug) — those land in 3f/3g.

* DESIGN.md updates the SkipAnalysisService row (drops the
  "deferred to index.ts" note) and adds the router row.

Stacked on phase3-pr4-shikimori; refs #102, #84.
