+++ DESIGN.md
@@ -42,6 +42,7 @@ Renderer (Vue)  --ipcRenderer.invoke-->  Preload (bridge)  --ipcMain.handle-->  
 | `src/main/ipc/downloaded-anime.ipc.ts` | Downloaded-anime router: `CHANNELS.{DOWNLOADED_ANIME_ADD, DOWNLOADED_ANIME_DELETE}` — delete cascades to all storage dirs, anime-cache entry, skip-detection cache + detections |
 | `src/main/ipc/shikimori.ipc.ts` | Shikimori router (Phase 3 slice 3d): all 16 `CHANNELS.SHIKIMORI_*` handlers — OAuth (`GET_AUTH_URL`/`EXCHANGE_CODE`/`LOGOUT`/`GET_USER`), rate read/write (`GET_RATE`/`UPDATE_RATE` with offline-queue fallback on network error), `GET_FRIENDS_RATES`/`GET_ANIME_RATES` (cache-first + background refresh)/`GET_ANIME_DETAILS`/`TRIGGER_DETAIL_PREFETCH`, sync surface (`GET_OFFLINE_QUEUE_LENGTH`/`GET_SYNC_STATUS`/`TRIGGER_SYNC`), `GET_FRIENDS_ACTIVITY`/`GET_CALENDAR`/`GET_RELATED`. Holds the `fetchAndCacheShikimoriRates`/`refreshShikimoriRatesInBackground` orchestration locally |
 | `src/main/services/shikimori-sync/index.ts` | `createShikimoriSyncService` — owns the Shikimori sync timer + offline-queue drain worker (`syncShikimoriQueue`, consolidate → drift-resolve → reconcile), the anime-details prefetcher (`prefetchShikimoriDetails`/`refreshShikimoriDetailsForMalId`), and the 5-minute calendar cache. All sync/prefetch/calendar module state (formerly free vars in `index.ts`) now lives in the service closure. Deps-injected (`store`, `broadcast` + 4 channel names) |
+| `src/main/ipc/skip-detector.ipc.ts` | Skip-detector router (Phase 3 slice 3e): `CHANNELS.SKIP_DETECTOR_{GET_DETECTIONS, GET_STATUS, CANCEL, CACHE_STATS, ANALYZE_SHOW, DETECT_STREAM, CANCEL_STREAM_DETECT, BACKFILL_ALL, QUEUE_STATUS}` + `DOWNLOAD_INJECT_CHAPTERS` (deferred from 3c — `fluent-ffmpeg` `-map_chapters` write over each downloaded MKV, with a fingerprint-cache key rewrite to survive the size/mtime change) |
 | `src/main/smotret-api.ts` | Smotret-Anime API client: search, anime/episode details, embed, subtitles, token validation |
 | `src/main/download-manager.ts` | Download queue, concurrent downloads, progress, ffmpeg merge, scan-merge |
 | `src/main/shikimori.ts` | Shikimori API client: OAuth, token refresh, user rates, anime list |
@@ -59,7 +60,7 @@ Renderer (Vue)  --ipcRenderer.invoke-->  Preload (bridge)  --ipcMain.handle-->  
 | `src/main/store/keys.ts` | `PERSISTED_STORE_KEYS` frozen tuple — rename guard, compile-time-asserted against `keyof STORE_DEFAULTS` in `index.ts` |
 | `src/main/store/migrate.ts` | Pure `migrateWatchProgressV2(store)` (one-shot `watchedAt` backfill) |
 | `src/main/services/anime-cache/index.ts` | `createAnimeCacheService` — `AnimeCacheEntry` CRUD + poster file cache + `cleanupStale` orphan sweep; deps-injected (`store`, `userDataDir`, `fetchPoster`, `isCachable`) |
-| `src/main/services/skip-analysis/index.ts` | `createSkipAnalysisService` — `normalizeDetections`, fingerprint cache `pruneCacheFor{Episode,Anime}` + `sweepFingerprintCache`, `dropDetectionsFor{Anime,Episode}`, `buildAutoSkipEpisodes`; deps-injected (`store`, `scanEpisodeFiles`, `sanitizeFilename`, `broadcast`). Auto-skip queue/drain machinery still lives in `index.ts` (deferred — entangled with stream-skip state). |
+| `src/main/services/skip-analysis/index.ts` | `createSkipAnalysisService` — leaf surface (`normalizeDetections`, fingerprint cache `pruneCacheFor{Episode,Anime}` + `sweepFingerprintCache`, `dropDetectionsFor{Anime,Episode}`, `buildAutoSkipEpisodes`) **plus** the orchestration folded in by Phase 3 slice 3e: full-show analysis single-flight (`runSkipAnalysis`/`getCurrentAnalysis`/`cancelCurrentAnalysis`), per-renderer stream detection (`runStreamSkipDetection`/`cancelStreamSkipDetection`), and the debounced auto-skip queue (`scheduleAutoSkipAnalysis`/`enqueueAutoSkipAnalysis`/`getAutoSkipQueueStatus`). All in-flight state (`currentSkipAnalysis`, the stream-detection map, the debounce map + queue) lives in the service closure. Deps-injected (`store`, `scanEpisodeFiles`, `sanitizeFilename`, `broadcast`, 2 channel names, `getFpcalcPath`/`getFfmpegPath`/`getFfprobePath`). |
 | `src/main/services/cold-storage/index.ts` | `createColdStorageService` — full hot/cold storage surface: path helpers (`getDownloadDir`, `getColdStorageDir`, `isAdvanced`, `dirsForScan`), `scanUsage` (hot/cold bucket walk → `StorageUsage`), disk mutators (`deleteEpisodeFiles`, `moveEpisodeToColdStorage`, `moveAllFilesToColdStorage`, `pruneDownloadedEpisode`), read-only checks (`episodeFileExists`, `episodeHasInProgressDownload`), and watched-cleanup orchestrator (`findCleanupCandidates`, `runWatchedCleanup` with confirm/force gating + per-run lock). Deps-injected (`store`, `userDataDir`, `sanitizeFilename`, `parseEpisodeFromFilename`, `scanEpisodeFiles`, `invalidateFileCache`, `broadcast` + 4 channel names). |
 | `src/main/streaming/codec-strings.ts` | Pure ffprobe→codec-string mappers: `avcCodecString`, `hevcCodecString`, `aacCodecString` (RFC 6381 / MSE-compatible). |
 | `src/main/streaming/index.ts` | `createStreamingService` — MSE session machinery: `probeMkvForMse`, `findKeyframeBefore`, `pickH264Encoder` (cached, vaapi → nvenc → qsv → libx264 dry-run-gated), `h264EncoderCandidates`, `spawnFfmpegForSession` (fragmented-MP4 pipe + backpressure-managed IPC chunks + ffmpeg stderr-throttled progress + HEVC `hvc1` tagging for MSE), session map (`registerSession`/`getSession`/`cleanupSession`/`cleanupAllSessions`), `tmpDir`, watermark constants. Deps-injected (`getFfmpegPath`/`getFfprobePath` live getters + 4 stream IPC channels). Player IPC handlers in `index.ts` consume the service primitives; their orchestration (handler bodies, HEVC consent modal flow, subtitle extraction) stays in `index.ts`. |
